+++ template-1/src/media/media.css
@@ -14,13 +14,13 @@
 }
 
 .media .title {
-  margin: 0 auto;
-  color: #1f172c;
+  /* margin: 0 auto; */
+  /* color: #1f172c;
   text-align: center;
   font-family: Satoshi;
   font-size: 24px;
   font-style: normal;
-  font-weight: 900;
+  font-weight: 900; */
   line-height: 21.92px; /* 91.333% */
   text-transform: uppercase;
 }
